@@ -6,9 +6,9 @@ import de.caritas.cob.consultingtypeservice.api.model.TopicMultilingualDTO;
 import de.caritas.cob.consultingtypeservice.api.service.TopicFeatureAuthorisationService;
 import de.caritas.cob.consultingtypeservice.api.service.TopicServiceFacade;
 import de.caritas.cob.consultingtypeservice.generated.api.controller.TopicadminApi;
-import io.swagger.annotations.Api;
+import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.validation.Valid;
 import java.util.List;
-import javax.validation.Valid;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -17,12 +17,14 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.util.CollectionUtils;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
-@Api(tags = "topic-controller")
 @Slf4j
+@Tag(name = "topic-controller")
 public class TopicAdminController implements TopicadminApi {
 
   private final @NonNull TopicServiceFacade topicServiceFacade;
@@ -33,6 +35,10 @@ public class TopicAdminController implements TopicadminApi {
 
   @Override
   @PreAuthorize("hasAuthority('AUTHORIZATION_CREATE_TOPIC')")
+  @RequestMapping(
+      method = RequestMethod.POST,
+      value = {"/topicadmin", "/topicadmin/"},
+      produces = {"application/json"})
   public ResponseEntity<TopicMultilingualDTO> createTopic(
       @Valid final TopicMultilingualDTO topicMultilingualDTO) {
     log.info("Creating topic by user {} ", authenticatedUser.getUsername());
@@ -52,6 +58,10 @@ public class TopicAdminController implements TopicadminApi {
 
   @Override
   @PreAuthorize("hasAuthority('AUTHORIZATION_GET_ALL_TOPICS')")
+  @RequestMapping(
+      method = RequestMethod.GET,
+      value = {"/topicadmin", "/topicadmin/"},
+      produces = {"application/json"})
   public ResponseEntity<List<TopicDTO>> getAllTopicsAsAdmin() {
     final var topics = topicServiceFacade.getAllTopics();
     return !CollectionUtils.isEmpty(topics)
